+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.0.3] - 16/04/2026
+
+* [feature] Constructive Progression Rate refined with event chain inspection (clearances & long passes in first 3 passes)
+* [refactor] Improved check_event_chain() to identify structured vs. defensive build-up
+
 ## [0.0.2] - 16/04/2026
 
 * [refactor] Move utility functions to src/helpers.py